+++ CHANGELOG.md
@@ -18,7 +18,7 @@ the code was deployed.
 - Added console function for the occupation detection timer as well as the timer count and timeout value in Debug Publishes (CU-8678t4ztg).
 - "doorMissedFrequently" field to JSON data submitted by brave sensors that is sent to /api/heartbeat (CU-860rk8v2a).
 - Sentry log in the case that doorMissedFrequently is true in posted data from brave sensor (CU-860rk8v2a).
-
+- GitHub Actions to deploy to AWS infrastructure on Dev and Staging (CU-860ra8f7q).
 ### Fixed
 
 - Wording in consoleFunctionTests.cpp when functions are called with 'e'.